chore: Update .gitignore and Makefile
This commit updates the .gitignore file to exclude the .vscode directory. It also modifies the Makefile to add a new test target called test_tmp. This target is used to run a specific test for the stage_12: Scanning: String literals. The test is executed using the main.out binary located in the dist directory.